+++ Release/_hijack_root/qml/Singletons/One_UserImageSelection.qml
@@ -8,11 +8,12 @@
  * - "Default"
  * - "MZ"
  * - "Koffin"
+ * - "Krypt"
 */
 //======================================================================================
 pragma Singleton
 import QtQuick 2.3
 
 QtObject {
-    readonly property string userSelection: "Default"
+    readonly property string userSelection: "MZ"
 }